# backend/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": " NRule - Rules Engine API",
        "version": "0.0.0",
        "description": "Rules Engine built for Gentera.\n t - Tenancy\n nr - NRule"
    },
    "paths": {
...
        "/nr/api/v1/t/{tid}/rules/{rid}": {
            "put": {
                "summary": "Update rule's attributes.",
                "tags": [
                    "Rules"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/TenantID"
                    },
                    {
                        "$ref": "#/parameters/RuleID"
                    },
                    {
                        "$ref": "#/parameters/IdType"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rule Updated"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "409": {
                        "description": "Rule already exists"
                    }
                }
            },
            "get": {
                "summary": "Read one rule.",
                "tags": [
                    "Rules"
                ],                
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/TenantID"
                    },
                    {
                        "$ref": "#/parameters/RuleID"
                    },
                    {
                        "$ref": "#/parameters/IdType"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rule Updated"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "409": {
                        "description": "Rule already exists"
                    }
                }
            }
        },
...
    },
    "parameters": {
        "TenantID":{
            "name": "tid",
            "description": "Tenant Identification",
            "in": "path",
            "type": "integer",
            "required": true
        },
        "RuleID":{
            "name": "rid",
            "description": "Rule Identification",
            "in": "path",
            "type": "string",
            "required": true
        },
...
        "IdType":{
            "name": "idType",
            "description": "Rule Identification Type.",
            "in": "query",
            "type": "string",
            "required": false,
            "enum": ["__external", "__internal"],
            "default": "__internal"
        },
...
    },
    "definitions": {
...
        "UpdateRule": {
            "type": "object",
            "required": ["strategy"],
            "properties": {
                "strategy": {
                    "description": "Type of strategy.",
                    "type": "string",
                    "enum": [
                        "ALL",
                        "BASE",
                        "EARLY"
                    ],
                    "default": "BASE"
                }
            }
        },
...
    }
}